@@ -3,6 +3,7 @@
 	"version": "1.0.0",
 	"description": "Weather forecasts.",
 	"license": "GPL-3.0-only",
+	"type": "module",
 	"author": {
 		"name": "Jenny Belanger",
 		"email": "mail@jennybelanger.com",
@@ -16,42 +17,43 @@
 		"build": "rm -rf build/ && node_modules/.bin/webpack --mode=production",
 		"lint": "yarn lint:css && yarn lint:js",
 		"lint:css": "node_modules/.bin/stylelint 'css/**/*.css'",
-		"lint:js": "node_modules/.bin/eslint . --ext .js",
+		"lint:js": "node_modules/.bin/eslint .",
 		"lint:php": "./vendor/bin/phpcs",
-		"start": "npx update-browserslist-db@latest && node_modules/.bin/webpack --watch"
+		"start": "node_modules/.bin/update-browserslist-db && node_modules/.bin/webpack --watch"
 	},
 	"dependencies": {
-		"@csstools/postcss-global-data": "^3.0.0",
-		"chart.js": "^4.5.0",
+		"@csstools/postcss-global-data": "^3.1.0",
+		"chart.js": "^4.5.1",
 		"chartjs-adapter-luxon": "^1.3.1",
 		"chartjs-plugin-annotation": "^3.1.0",
-		"luxon": "^3.7.1",
+		"luxon": "^3.7.2",
 		"normalize.css": "^8.0.1"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.26.0",
-		"@stylistic/stylelint-plugin": "^3.1.1",
-		"babel-loader": "^9.2.1",
-		"browser-sync": "^3.0.3",
-		"browser-sync-webpack-plugin": "^2.3.0",
-		"copy-webpack-plugin": "^12.0.2",
+		"@eslint/js": "^9.39.2",
+		"@stylistic/eslint-plugin": "^5.6.1",
+		"@stylistic/stylelint-plugin": "^4.0.0",
+		"babel-loader": "^10.0.0",
+		"browser-sync": "^3.0.4",
+		"browser-sync-webpack-plugin": "^2.4.0",
+		"copy-webpack-plugin": "^13.0.1",
 		"css-loader": "^7.1.2",
-		"css-minimizer-webpack-plugin": "^7.0.2",
-		"dotenv": "^16.4.7",
-		"eslint": "^8.57.1",
-		"eslint-config-airbnb-base": "^15.0.0",
-		"eslint-plugin-import": "^2.31.0",
-		"mini-css-extract-plugin": "^2.9.2",
-		"postcss": "^8.4.49",
-		"postcss-loader": "^8.1.1",
-		"postcss-preset-env": "^10.1.2",
-		"stylelint": "^16.18.0",
-		"stylelint-config-standard": "^38.0.0",
-		"stylelint-order": "^6.0.4",
-		"terser-webpack-plugin": "^5.3.11",
-		"webpack": "^5.97.1",
+		"css-minimizer-webpack-plugin": "^7.0.4",
+		"eslint": "^9.39.2",
+		"eslint-plugin-import": "^2.32.0",
+		"globals": "^16.5.0",
+		"mini-css-extract-plugin": "^2.9.4",
+		"postcss": "^8.5.6",
+		"postcss-loader": "^8.2.0",
+		"postcss-preset-env": "^10.5.0",
+		"stylelint": "^16.26.1",
+		"stylelint-config-standard": "^39.0.1",
+		"stylelint-order": "^7.0.0",
+		"terser-webpack-plugin": "^5.3.16",
+		"webpack": "^5.104.1",
 		"webpack-cli": "^6.0.1",
-		"webpack-manifest-plugin": "^5.0.0"
+		"webpack-manifest-plugin": "^6.0.1"
 	},
 	"browserslist": [
 		"defaults",
